third_party/WebKit/LayoutTests/typedcssom/cssTranslation.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

var simpleLength = new CSSUnitValue(0, "px");
var decimalLength = new CSSUnitValue(1.1, "px");
var negativeLength = new CSSUnitValue(-2.2, "em");
// var calcLengthPx = new CSSCalcValue({px: 1});
// var calcLength = new CSSCalcValue({px: 1, em: -2.2});

var simplePercent = new CSSUnitValue(10, "percent");
// var calcPercent = new CSSCalcValue({px: 1, percent: 2.2});

var testParams = [
  // 2D CSSTranslation Transform Components
  {
    input: new CSSTranslation(simpleLength, simpleLength),
    x: simpleLength, y: simpleLength,
    is2D: true,
    cssText: "translate(0px, 0px)"
  },
  {
    input: new CSSTranslation(decimalLength, negativeLength),
    x: decimalLength, y: negativeLength,
    is2D: true,
    cssText: "translate(1.1px, -2.2em)"
  },
  // {input: new CSSTranslation(negativeLength, calcLengthPx),
  //   x: negativeLength, y: calcLengthPx, is2D: true},
  // {input: new CSSTranslation(calcLengthPx, negativeLength),
  //   x: calcLengthPx, y: negativeLength, is2D: true},
  // {input: new CSSTranslation(calcLengthPx, calcLength),
  //   x: calcLengthPx, y: calcLength, is2D: true},
  {
    input: new CSSTranslation(simplePercent, simpleLength),
    x: simplePercent, y: simpleLength,
    is2D: true,
    cssText: "translate(10%, 0px)"
  },
  // {input: new CSSTranslation(calcLengthPx, simplePercent),
  //   x: calcLengthPx, y: simplePercent, is2D: true},
  // {input: new CSSTranslation(calcPercent, calcLength),
  //   x: calcPercent, y: calcLength, is2D: true},
  // {input: new CSSTranslation(simplePercent, calcPercent),
  //   x: simplePercent, y: calcPercent, is2D: true},

  // 3D CSSTranslation Transform Components
  {
    input: new CSSTranslation(simpleLength, simpleLength, simpleLength),
    x: simpleLength, y: simpleLength, z: simpleLength,
    is2D: false,
    cssText: "translate3d(0px, 0px, 0px)"
  },
  {
    input: new CSSTranslation(simpleLength, decimalLength, negativeLength),
    x: simpleLength, y: decimalLength, z: negativeLength,
    is2D: false,
    cssText: "translate3d(0px, 1.1px, -2.2em)"
  },
  // {input: new CSSTranslation(simpleLength, simpleLength, calcLengthPx),
  //   x: simpleLength, y: simpleLength, z: calcLengthPx, is2D: false},
  // {input: new CSSTranslation(calcLengthPx, calcLength, calcLength),
  //   x: calcLengthPx, y: calcLength, z: calcLength, is2D: false},
  {
    input: new CSSTranslation(simplePercent, decimalLength, simpleLength),
    x: simplePercent, y: decimalLength, z: simpleLength,
    is2D: false,
    cssText: "translate3d(10%, 1.1px, 0px)"
  },
  // {input: new CSSTranslation(simpleLength, calcPercent, decimalLength),
  //   x: simpleLength, y: calcPercent, z: decimalLength, is2D: false},
  // {input: new CSSTranslation(calcPercent, simplePercent, calcLength),
  //   x: calcPercent, y: simplePercent, z: calcLength, is2D: false}
];

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.x, params.x);
    assert_equals(params.input.y, params.y);
    if (params.is2D) {
      assert_equals(params.input.z, null);
    } else {
      assert_equals(params.input.z, params.z);
    }
  }, "x, y, and z values are correct for " + params.cssText);
}

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.is2D(), params.is2D);
  }, "is2D value is correct for " + params.cssText);
}

for (let params of testParams) {
  test(() => {
    assert_equals(params.input.toString(), params.cssText);
  }, "toString value is correct for " + params.cssText);
}

test(() => {
  assert_throws(new TypeError(), () => {
    new CSSTranslation(simpleLength, simpleLength, simplePercent);
  });
  // assert_throws(null, () => {
  //  new CSSTranslation(simpleLength, simpleLength, calcPercent);
  // });
  assert_throws(new TypeError(), () => {
    new CSSTranslation(simplePercent, simplePercent, simplePercent);
  });
}, "Constructor throws when z component contains percent.");

test(() => {
  assert_throws(new TypeError(), () => { new CSSTranslation(); });
  assert_throws(new TypeError(), () => { new CSSTranslation(simpleLength); });
}, "Invalid number of arguments to constructor throws an exception.");

</script>
